test: add comprehensive integration tests for Fast First Pass workflow
- Add test_fast_first_pass.py with 9 comprehensive integration tests
- Test dual-pass vs single-pass reading workflows
- Verify model configuration (llama3.1:8b + qwen3:8b) integration
- Test Spanish document processing with dual-pass approach
- Validate SPECS v2.0 JSON output structure compliance
- Test markdown cleaning integration (titles + sections)
- Performance comparison between dual-pass and single-pass
- Fix avg_summary_length type from int to float in CognitiveKnowledge model
- All 9 integration tests passing + 129 total tests passing

+++ src/cognitive_reader/models/document.py
@@ -77,7 +77,7 @@ class CognitiveKnowledge(BaseModel):
 
     # Document Statistics (useful for apps consuming the data)
     total_sections: int = Field(description="Total number of sections processed")
-    avg_summary_length: int = Field(description="Average summary length in characters")
+    avg_summary_length: float = Field(description="Average summary length in characters")
     total_concepts: int = Field(description="Total number of concepts identified")
 
     def get_summary_by_id(self, section_id: str) -> SectionSummary | None:
